Tighten up floating point code
c and sigma are stored into TCutData which is typedef'd to
float which is why they are down converted before storing.
So there is no need for double precision math in the first
place! We can set both of those to float and lose the
conversion. This saves conversion instructions plus the
fdiv's will be faster. Also c is not used so just compute
1/c which saves another divide.

No loss in fidelity, no need to change tolerances.